Fix items vanishing through end portal
If the Paper configuration option "keep-spawn-loaded" is set to false,
items entering the overworld from the end will spawn at Y = 0.

This is due to logic in the getHighestBlockYAt method in World.java
only searching the heightmap if the chunk is loaded.

Quickly loading the exact world spawn chunk before searching the
heightmap resolves the issue without having to load all spawn chunks.