Disable allowListing before received from client
The client does not send the packet needed to received this information until a little later, which can cause a condition where
despite a player having disabled listing, they are able to be seen for a brief moment. This causes the player to be listed as an Anonymous
Player until the actual configuration value is received from the client.